+++ packages/react-tinacms-editor/src/components/MenuHelpers/MenuButton.tsx
@@ -23,6 +23,7 @@ const MenuItem = css`
 `
 
 export const MenuButton = styled.button<{
+  title?: string
   active?: boolean
   disabled?: boolean
   ref?: any