+++ packages/@tinacms/react-toolbar/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.27.1](https://github.com/tinacms/tinacms/compare/v0.27.0...v0.27.1) (2020-08-10)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+# [0.26.0](https://github.com/tinacms/tinacms/compare/v0.25.0...v0.26.0) (2020-08-03)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
 # [0.4.0-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.3.1...@tinacms/react-toolbar@0.4.0-alpha.0) (2020-07-15)